pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "telegram-navigator"
version = "0.1.0"
description = "AI-powered table-of-contents bot for Telegram channels"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "python-telegram-bot>=21.0",
    "telethon==1.36.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
supabase = ["supabase>=2.0.0"]
embeddings = ["openai>=1.0.0", "sqlite-vec>=0.1.6"]
dev = ["ruff>=0.4.0"]
all = ["telegram-navigator[supabase,embeddings]"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
